pyproject.toml
[tool.poetry]
name = "stat_arb_prod"
version = "0.1.0"
description = "Production-grade statistical arbitrage trading system"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.0"
numpy = "^1.24.0"
psycopg2-binary = "^2.9.0"
SQLAlchemy = "^2.0.0"
statsmodels = "^0.14.0"
scipy = "^1.10.0"
scikit-learn = "^1.3.0"
mlflow = "^2.8.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
matplotlib = "^3.6.0"
seaborn = "^0.12.0"
plotly = "^5.17.0"
gunicorn = "^20.0.0"
uvicorn = "^0.20.0"
redis = "^4.0.0"
numba = "^0.58.0"
joblib = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
